+++ src/messages/main.py
@@ -1,10 +1,26 @@
 import logging.config
 import os
 import uuid
+import sentry_sdk
 
 from common.logging_config import get_logging_configuration
 from messages.ws import MessageInterface
 from uhopper.utils.mqtt.handler import MqttPublishHandler
+from sentry_sdk.integrations.logging import LoggingIntegration
+from sentry_sdk.integrations.flask import FlaskIntegration
+
+sentry_logging = LoggingIntegration(
+    level=logging.INFO,  # Capture info and above as breadcrumbs
+    event_level=logging.ERROR  # Send errors as events
+)
+
+sentry_sdk.init(
+    integrations=[FlaskIntegration()],
+    # Set traces_sample_rate to 1.0 to capture 100%
+    # of transactions for performance monitoring.
+    # We recommend adjusting this value in production.
+    traces_sample_rate=1.0
+)
 
 bot_id = os.getenv("BOT_ID")